Young

+++ src/repositories/post.repository.js
@@ -1,5 +1,7 @@
 import Post from "../models/post.js";
+import Comment from "../models/comment.js";
 import User from "../models/user.js";
+import Like from "../models/like.js";
 
 class Postrepositories {
     // 만들어야 할꺼
@@ -9,6 +11,12 @@ class Postrepositories {
     // 게시판 상세보기 = postviewdetail
     // 커뮤니티 수정 = postupdate
     // 커뮤니티 삭제 = postdelete
+
+    finduser = async (UserId) => {
+        const user = await User.findOne({where:{UserId},raw:true})
+
+        return user;
+    };
 
     postview = async () => {
         const list = await Post.findAll({
@@ -26,7 +34,7 @@ class Postrepositories {
         const post = await Post.findOne({
             where: { id: postId },
             include: [{ model: User, attributes: ["nickname"] }],
-            attributes: { exclude: ["nickname", "UserId"] },
+            attributes: { exclude: ["nickname"] },
             raw: true
         });
 
@@ -39,17 +47,35 @@ class Postrepositories {
         return;
     };
 
-    postupdate = async (title, content, postId) => {
-        await Post.update({ title, content }, { where: { id: postId } });
+    postupdate = async (title, content, postId, UserId) => {
+        await Post.update({ title, content }, { where: { id: postId, UserId } });
 
         return;
     };
 
-    postdelete = async (postId) => {
-        await Post.destroy({ where: { id: postId } });
+    postdelete = async (postId, UserId) => {
+        await Post.destroy({ where: { id: postId, UserId } });
 
         return;
     };
+
+    postLike = async(postId, userId) => {
+        const likeCreate = await Like.create({
+            PostId : postId,
+            UserId : userId,
+            like : true,
+        })
+
+        return likeCreate;
+    }
+
+    postLikeDelete = async(postId, userId) => {
+        const LikeDelete = await Like.destroy({
+            where : {PostId: postId, UserId : userId}
+        });
+
+        return LikeDelete
+    }
 }
 
 export default new Postrepositories;
